Re-structuring and renaming files in Muon script directory
