+++ scripts/Muon/GUI/Common/plot_widget/external_plotting/external_plotting_model.py
