+++ scripts/Muon/GUI/Common/plot_widget/plot_widget_view_interface.py
@@ -15,7 +15,7 @@ class PlottingWidgetViewMeta(type(QtWidgets.QWidget), ABCMeta):
     pass
 
 
-class PlottingWidgetViewInterface(metaclass=PlottingWidgetViewMeta):
+class PlotWidgetViewInterface(metaclass=PlottingWidgetViewMeta):
 
     @abstractmethod
     def setup_plot_type_options(self, options: List[str]):
