+++ scripts/Muon/GUI/Common/plot_widget/plotting_canvas/plotting_canvas_model.py
