+++ scripts/Muon/GUI/Common/plot_widget/plotting_widget_view.ui
