+++ scripts/Muon/GUI/Common/plotting_dock_widget/plotting_dock_widget.py
@@ -4,7 +4,7 @@
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
-from qtpy import QtWidgets, QtCore
+from qtpy import QtWidgets, QtCore, QtGui
 
 
 class PlottingDockWidget(QtWidgets.QDockWidget):
@@ -19,11 +19,3 @@ class PlottingDockWidget(QtWidgets.QDockWidget):
         self.dockable_plot_widget = plotting_widget
         # set the widget for this dockable plot window
         self.setWidget(self.dockable_plot_widget)
-
-        self.dockLocationChanged.connect(self.window_dock_status_changed)
-
-    # this function makes the plot rescale after being docked or un-docked,
-    # to avoid the x labels, y labels and title disappearing upon this action
-    def window_dock_status_changed(self, area):
-        pass
-        #self.dockable_plot_widget.get_fig().tight_layout()
