+++ scripts/Muon/GUI/FrequencyDomainAnalysis/frequency_domain_analysis_2.py
@@ -31,7 +31,7 @@ from Muon.GUI.MuonAnalysis.load_widget.load_widget import LoadWidget
 from Muon.GUI.Common.phase_table_widget.phase_table_widget import PhaseTabWidget
 from Muon.GUI.Common.results_tab_widget.results_tab_widget import ResultsTabWidget
 from Muon.GUI.Common.fitting_tab_widget.fitting_tab_widget import FittingTabWidget
-from Muon.GUI.Common.plotting_widget.plotting_widget import PlottingWidget
+from Muon.GUI.Common.plot_widget.plot_widget import PlotWidget
 from Muon.GUI.Common.plotting_dock_widget.plotting_dock_widget import PlottingDockWidget
 
 SUPPORTED_FACILITIES = ["ISIS", "SmuS"]
@@ -88,9 +88,9 @@ class FrequencyAnalysisGui(QtWidgets.QMainWindow):
             workspace_suffix=' FD', frequency_context=self.frequency_context)
 
         # create the dockable widget
-        self.dockable_plot_widget = PlottingWidget(self.context)
+        self.plot_widget = PlotWidget(self.context)
         self.dockable_plot_widget_window = PlottingDockWidget(parent=self,
-                                                              plotting_widget=self.dockable_plot_widget.view)
+                                                              plotting_widget=self.plot_widget.view)
         self.dockable_plot_widget_window.setMinimumWidth(575)
 
         # Add dock widget to main Muon analysis window
@@ -156,7 +156,7 @@ class FrequencyAnalysisGui(QtWidgets.QMainWindow):
         self.transform.new_data_observer(
             self.fitting_tab.fitting_tab_presenter.input_workspace_observer)
         self.transform.new_data_observer(
-            self.dockable_plot_widget.presenter.input_workspace_observer)
+            self.plot_widget.presenter.input_workspace_observer)
 
         self.context.data_context.message_notifier.add_subscriber(
             self.grouping_tab_widget.group_tab_presenter.message_observer)
@@ -192,7 +192,7 @@ class FrequencyAnalysisGui(QtWidgets.QMainWindow):
             self.phase_tab.phase_table_presenter.run_change_observer)
 
         self.grouping_tab_widget.group_tab_presenter.calculation_finished_notifier.add_subscriber(
-            self.dockable_plot_widget.presenter.input_workspace_observer)
+            self.plot_widget.presenter.input_workspace_observer)
 
     def setup_gui_variable_observers(self):
         self.context.gui_context.gui_variables_notifier.add_subscriber(
@@ -209,15 +209,15 @@ class FrequencyAnalysisGui(QtWidgets.QMainWindow):
             self.fitting_tab.fitting_tab_presenter.selected_group_pair_observer)
 
         self.grouping_tab_widget.pairing_table_widget.selected_pair_changed_notifier.add_subscriber(
-            self.dockable_plot_widget.presenter.added_group_or_pair_observer)
+            self.plot_widget.presenter.added_group_or_pair_observer)
 
         self.grouping_tab_widget.grouping_table_widget.selected_group_changed_notifier.add_subscriber(
             self.fitting_tab.fitting_tab_presenter.selected_group_pair_observer)
 
         self.grouping_tab_widget.grouping_table_widget.selected_group_changed_notifier.add_subscriber(
-            self.dockable_plot_widget.presenter.added_group_or_pair_observer)
+            self.plot_widget.presenter.added_group_or_pair_observer)
 
-        self.dockable_plot_widget.presenter.plot_type_changed_notifier.add_subscriber(
+        self.plot_widget.presenter.plot_type_changed_notifier.add_subscriber(
             self.fitting_tab.fitting_tab_presenter.selected_plot_type_observer)
 
     def setup_grouping_changed_observers(self):
@@ -291,10 +291,10 @@ class FrequencyAnalysisGui(QtWidgets.QMainWindow):
             self.fitting_tab.fitting_tab_presenter.input_workspace_observer)
 
         self.grouping_tab_widget.group_tab_presenter.calculation_finished_notifier.add_subscriber(
-            self.dockable_plot_widget.presenter.input_workspace_observer)
+            self.plot_widget.presenter.input_workspace_observer)
 
         self.grouping_tab_widget.group_tab_presenter.calculation_finished_notifier.add_subscriber(
-            self.dockable_plot_widget.presenter.rebin_options_set_observer)
+            self.plot_widget.presenter.rebin_options_set_observer)
 
     def setup_phase_quad_changed_notifier(self):
         self.phase_tab.phase_table_presenter.phase_quad_calculation_complete_nofifier.add_subscriber(
@@ -309,12 +309,12 @@ class FrequencyAnalysisGui(QtWidgets.QMainWindow):
         self.fitting_context.new_fit_results_notifier.add_subscriber(
             self.results_tab.results_tab_presenter.new_fit_performed_observer)
 
-        self.fitting_context.new_fit_plotting_notifier.add_subscriber(self.dockable_plot_widget.presenter.fit_observer)
+        self.fitting_context.new_fit_plotting_notifier.add_subscriber(self.plot_widget.presenter.fit_observer)
 
-        self.fitting_context.fit_removed_notifier.add_subscriber(self.dockable_plot_widget.presenter.
+        self.fitting_context.fit_removed_notifier.add_subscriber(self.plot_widget.presenter.
                                                                  fit_removed_observer)
 
-        self.fitting_context.plot_guess_notifier.add_subscriber(self.dockable_plot_widget.presenter.plot_guess_observer)
+        self.fitting_context.plot_guess_notifier.add_subscriber(self.plot_widget.presenter.plot_guess_observer)
 
     def closeEvent(self, event):
         self.tabs.closeEvent(event)
